Update README.md
I made some changes to the readme since I had a problem with it at first ( im on linux ) and I wanted to adjust it so people could have a easier time doing the task at hand. I also formatted it a bit since the OS was a bit small compared to the instructions. Hope it forks, thanks.